refactor: extract Car class to separate file
