+++ src/main/kotlin/racingcar/Application.kt
@@ -1,24 +1,16 @@
 package racingcar
+
 import camp.nextstep.edu.missionutils.Console
-import camp.nextstep.edu.missionutils.Randoms
 
 fun main() {
     println("Enter the names of the cars(comma-separated):")
     val input = Console.readLine()
-    val splitNames = input.split(",", ignoreCase = true, limit = 0) //delimiter can be "," or ", "
+    val splitNames = input.split(",")
     val names = validateCarNames(splitNames)
 
     println("How many rounds will be played?")
     val rounds = Console.readLine()
     val numberOfRounds: Int = validateRounds(rounds)
-
-    //println(numberOfRounds)
-
-//    for(i in splitNames.indices){
-//        val carName = splitNames[i].removeWhitespaces()
-//        names.add(carName)
-//    }
-    //println(names)
 
     val cars = names.map { Car(it) }
     val race = Race(cars, numberOfRounds)
@@ -78,29 +70,3 @@ class Race (private val cars: List<Car>, private val numberOfRounds: Int){
     }
 }
 
-class Car(val carName:String) {
-    private var speed :Int = 0
-    private val movement = mutableListOf<Char>()
-    var displacement = 0
-
-    private fun getSpeed() {
-        speed = Randoms.pickNumberInRange(0, 9)
-    }
-
-    fun moveCar() {
-        getSpeed()
-        if (speed >= 4) {
-            movement.add('-')
-            displacement++
-        }
-        println("$carName : ${movement.joinToString("")}")
-    }
-
-    //for testing the moveCar()
-    fun forceMove(carSpeed:Int) {
-        if (carSpeed >= 4) {
-            movement.add('-')
-            displacement++
-        }
-    }
-}
